Adjusted scripting convention.
Adjusted conventions for enemy scripts. Renamed EnemyMovement to EnemyPatroll and EnemyAttack to EnemyAttackInRange.
